refactor(core): rename picoschema package and drop legacy routes (#1116)
