+++ test-int/test_picoschema/helpers.py
@@ -9,7 +9,7 @@ from pathlib import Path
 
 import yaml
 
-from basic_memory.schema.inference import ObservationData, RelationData
+from basic_memory.picoschema.inference import ObservationData, RelationData
 
 
 # --- Fixture Paths ---
