+++ tests/api/v2/test_project_router.py
@@ -628,71 +628,3 @@ async def test_resolve_project_empty_identifier(client: AsyncClient, v2_projects
     response = await client.post(f"{v2_projects_url}/resolve", json=resolve_data)
 
     assert response.status_code == 422  # Validation error
-
-
-# --- Legacy v1 compatibility tests ---
-
-
-@pytest.mark.asyncio
-async def test_legacy_v1_list_projects_endpoint(client: AsyncClient, test_project: Project):
-    """Test that the legacy /projects/projects endpoint still works for older CLI versions.
-
-    This endpoint was removed when we migrated to v2 but older versions of
-    basic-memory-cloud CLI still call it for `bm project list`.
-
-    Note: The route must be without trailing slash to avoid 307 redirects
-    that the cloud proxy doesn't follow.
-    """
-    # The legacy v1 endpoint was at /projects/projects (no trailing slash)
-    response = await client.get("/projects/projects")
-
-    assert response.status_code == 200
-    data = response.json()
-    assert "projects" in data
-
-    # default_project must be populated, not null
-    assert data["default_project"] == test_project.name
-
-    project_names = [p["name"] for p in data["projects"]]
-    assert test_project.name in project_names
-
-
-@pytest.mark.asyncio
-async def test_legacy_v1_add_project_endpoint(client: AsyncClient, test_project: Project):
-    """Test that the legacy POST /projects/projects endpoint still works for older CLI versions.
-
-    Older versions of basic-memory-cloud CLI call POST /projects/projects to add projects.
-    The legacy route must proxy through to the same handler as v2.
-
-    Uses the existing test project name+path to exercise the idempotent path (200 OK),
-    which proves the route is connected without needing a full config manager.
-    """
-    response = await client.post(
-        "/projects/projects",
-        json={
-            "name": test_project.name,
-            "path": test_project.path,
-            "set_default": False,
-        },
-    )
-
-    # Idempotent: same name + same path returns 200
-    assert response.status_code == 200
-    data = response.json()
-    assert data["status"] == "success"
-    assert test_project.name in data["message"]
-
-
-@pytest.mark.asyncio
-async def test_legacy_v1_sync_config_endpoint(client: AsyncClient):
-    """Test that the legacy POST /projects/config/sync endpoint still works for older CLI versions.
-
-    Older versions of basic-memory-cloud CLI call POST /projects/config/sync to synchronize
-    projects between config file and database. The route must be reachable.
-    """
-    response = await client.post("/projects/config/sync")
-
-    # The handler synchronizes config ↔ DB; should succeed in test environment
-    assert response.status_code == 200
-    data = response.json()
-    assert data["status"] == "success"
